users_service/requirements.txt
django==4.2
djangorestframework==3.15
gunicorn==20.1.0
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger
requests